feat: only allow player registration when looking for players
